@Me decorator made. Ready for testing
