+++ src/controllers/user/user.controller.ts
@@ -6,6 +6,7 @@ import { UpdateUserDto } from '../../dto/user/update-user-dto';
 import { Auth } from '../../decorators/auth.decorator';
 import { AuthorizationService } from '../../services/authorization/authorization.service';
 import { SummaryUserDto } from '../../dto/user/summary-user-dto';
+import { Me } from '../../decorators/me.decorator';
 
 @ApiUseTags('User')
 @Controller('user')
@@ -23,32 +24,8 @@ export class UserController {
     })
     @ApiResponse({ status: 200, description: 'Found you', type: User })
     @ApiResponse({ status: 500, description: 'Internal server error...' })
-    async readMe(@Req() request: any) {
-        if (!request.headers.cookie) {
-            throw new UnauthorizedException('Geen koekje gevonden in je request... Zorg ervoor dat deze meegestuurd wordt met iedere request!');
-        }
-        const list: any = {};
-        const rc = request.headers.cookie;
-        if (rc) {
-            rc.split(';').forEach((cookie) => {
-                const parts = cookie.split('=');
-                list[parts.shift().trim()] = decodeURI(parts.join('='));
-            });
-        }
-
-        const auth = list.auth;
-        if (!auth) {
-            throw new UnauthorizedException('No authorizatie koekje gevonden... Zorg ervoor dat deze meegestuurd wordt met iedere request!');
-        }
-        const verifytoken = this.authService.verifyJWT(auth);
-        if (verifytoken) {
-            const decodedJWT = this.authService.decodeJWT(auth);
-            const user: User = await this.userService.readOne(decodedJWT.userId, decodedJWT.email);
-            return user;
-
-        } else {
-            throw new UnauthorizedException('Token incorrect of verlopen...');
-        }
+    async readMe(@Me() me: Promise<User>) {
+        console.log(me)
     }
 
     @Get('/:id')
